requirements.txt
PyQt6>=6.6.0
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-ApplicationServices>=10.0
flask>=3.0.0
pyinstaller>=6.0.0
